# Imagehash performance improvement and added initial support for blurhash