--- a/imagehash/transforms/etcs.go
+++ b/imagehash/transforms/etcs.go
@@ -30,10 +30,20 @@ func MedianOfPixels(pixels []float64) float64 {
 	return v
 }
 
-// MedianOfPixelsFast function returns a median value of pixels.
+// MedianOfPixelsHash64 function returns a median value of pixels.
 // It uses quick selection algorithm.
-func MedianOfPixelsFast(pixels []float64) float64 {
+func MedianOfPixels64(pixels []float64) float64 {
 	tmp := [64]float64{}
+	copy(tmp[:], pixels)
+	l := len(tmp)
+	pos := l / 2
+	return quickSelectMedian(tmp[:], 0, l-1, pos)
+}
+
+// MedianOfPixelsHash256 function returns a median value of pixels.
+// It uses quick selection algorithm.
+func MedianOfPixels256(pixels []float64) float64 {
+	tmp := [256]float64{}
 	copy(tmp[:], pixels)
 	l := len(tmp)
 	pos := l / 2