… go.sum
… imagehash/imagehash_gen_test.go
… imagehash/imagehash_test.go
… imagehash/transforms/dct_test.go
… imagehash/transforms/pixels.go
… imagehash/transforms/static.go