### Imagehash Changes
- Added support for PHash256
- Renamed Phash to PHash64, with alias for backward compatibility
- Increased Imagehash performance for PerceptionHash (PHash64 and PHash256)
- Added initial support for encoding a blurhash
- Improved tests for Imagehash

### Imagehash Performance
```
name                      old time/op    new time/op    delta
PHash64/Fast-12             65.2µs ± 1%    40.8µs ± 0%  -37.48%  (p=0.000 n=85+16)
PHash64/Fast-Parallel-12    10.6µs ± 1%     6.5µs ± 0%  -38.05%  (p=0.000 n=94+19)
PHash256/Fast-12            762µs ± 1%
PHash256/Fast-Parallel-12   122µs ± 0%
```